Frontend browser prepared
We have created all necessary modules for a functional frontend browser that users can use. Tab buttons, address bar, and the data store that saves the data about the opened or closed browser tab, including the url it was on before the user switches to, or adds another tab on the browser page. We just have to create the backend logic for showing and hiding the browser itself